backend/requirements.txt
# FastAPI 관련 의존성
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# MongoDB 관련 의존성
pymongo==4.6.0
motor

# 인증 관련 의존성
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# AI/ML 관련 의존성
openai-whisper==20231117
torch
numpy

# 유틸리티 의존성
python-dotenv==1.0.0
requests>=2.32.2
aiofiles
<<<<<<< HEAD


=======
httpx==0.25.2
>>>>>>> origin/main
